Separate the client components
